update collections design
