+++ src/app/views/common/lazy-loader/component-registry/index.tsx
@@ -6,7 +6,7 @@ import LazyStatusMessages from  '../../../app-sections/StatusMessages';
 import LazyResponseHeaders from  '../../../query-response/headers/ResponseHeaders';
 import LazyGraphToolkit from  '../../../query-response/graph-toolkit/GraphToolkit';
 import LazyCopyButton from  '../../copy-button/CopyButtonV9';
-import LazyAuth from  '../../../query-runner/request/auth/AuthV9';
+import LazyAuth from  '../../../query-runner/request/auth/Auth';
 import LazyRequestHeaders from  '../../../query-runner/request/headers/RequestHeaders';
 import LazyHistory from  '../../../sidebar/history/History';
 import LazyResourceExplorer from  '../../../sidebar/resource-explorer/ResourceExplorer';
