+++ src/app/views/query-runner/request/permissions/PermissionItem.tsx
@@ -1,9 +1,9 @@
 import {
-  DefaultButton, FontSizes, IColumn, IIconProps, IconButton, Label,
-  PrimaryButton, TooltipHost, getId, getTheme
-} from '@fluentui/react';
-
-import { AppDispatch, useAppDispatch, useAppSelector } from '../../../../../store';
+  Button,
+  Tooltip,
+  Label
+} from '@fluentui/react-components';
+import { useAppDispatch, useAppSelector } from '../../../../../store';
 import { IPermission, IPermissionGrant } from '../../../../../types/permissions';
 import { revokeScopes } from '../../../../services/actions/revoke-scopes.action';
 import { REVOKING_PERMISSIONS_REQUIRED_SCOPES } from '../../../../services/graph-constants';
@@ -11,34 +11,25 @@ import { consentToScopes } from '../../../../services/slices/auth.slice';
 import { getAllPrincipalGrant, getSinglePrincipalGrant } from '../../../../services/slices/permission-grants.slice';
 import { translateMessage } from '../../../../utils/translate-messages';
 import { PermissionConsentType } from './ConsentType';
-import { permissionStyles } from './Permission.styles';
+import { InfoRegular } from '@fluentui/react-icons';
+import permissionStyles from './Permission.styles';
 
 interface PermissionItemProps {
-  item: IPermission; index: number; column: IColumn | undefined;
+  item: IPermission;
+  index: number;
+  column: { key: string; fieldName?: string } | undefined;
 }
 
-const buttonIcon: IIconProps = {
-  iconName: 'Info',
-  style: {
-    position: 'relative',
-    top: '1px',
-    left: '6px'
-  }
-};
-
-const infoIcon: IIconProps = {
-  iconName: 'Info'
-};
-
 const PermissionItem = (props: PermissionItemProps): JSX.Element | null => {
-  const theme = getTheme();
-  const dispatch: AppDispatch = useAppDispatch();
-  const hostId: string = getId('tooltipHost');
-  const { scopes, auth: { consentedScopes }, profile: { user }, permissionGrants } = useAppSelector((state) => state);
+  const dispatch = useAppDispatch();
+  const scopes = useAppSelector((state) => state.scopes);
+  const consentedScopes  = useAppSelector((state) => state.auth.consentedScopes);
+  const user  = useAppSelector((state) => state.profile.user);
+  const permissionGrants = useAppSelector((state) => state.permissionGrants);
   const { item, column } = props;
-  const consented = !!item.consented;
+  const styles = permissionStyles();
 
-  const { adminLabelStyles, consentButtonStyles } = permissionStyles(theme);
+  const consented = !!item.consented;
 
   const handleConsent = async (permission: IPermission): Promise<void> => {
     const consentScopes = [permission.value];
@@ -47,19 +38,20 @@ const PermissionItem = (props: PermissionItemProps): JSX.Element | null => {
 
   const getAllPrincipalPermissions = (tenantWidePermissionsGrant: IPermissionGrant[]): string[] => {
     const allPrincipalPermissions = tenantWidePermissionsGrant.find((permission: IPermissionGrant) =>
-      permission.consentType.toLowerCase() === 'AllPrincipals'.toLowerCase());
+      permission.consentType.toLowerCase() === 'allprincipals'
+    );
     return allPrincipalPermissions ? allPrincipalPermissions.scope.split(' ') : [];
-  }
+  };
 
   const userHasRequiredPermissions = (): boolean => {
     if (permissionGrants && permissionGrants.permissions && permissionGrants.permissions.length > 0) {
       const allPrincipalPermissions = getAllPrincipalPermissions(permissionGrants.permissions);
       const principalAndAllPrincipalPermissions = [...allPrincipalPermissions, ...consentedScopes];
       const requiredPermissions = REVOKING_PERMISSIONS_REQUIRED_SCOPES.split(' ');
-      return requiredPermissions.every(scope => principalAndAllPrincipalPermissions.includes(scope));
+      return requiredPermissions.every((scope) => principalAndAllPrincipalPermissions.includes(scope));
     }
     return false;
-  }
+  };
 
   const ConsentTypeProperty = (): JSX.Element | null => {
     if (scopes && consented && user?.id) {
@@ -68,13 +60,16 @@ const PermissionItem = (props: PermissionItemProps): JSX.Element | null => {
       const singlePrincipalPermissions: string[] = getSinglePrincipalGrant(tenantWideGrant, user?.id);
       const tenantGrantFetchPending = permissionGrants.pending;
       const consentTypeProperties = {
-        item, allPrincipalPermissions, singlePrincipalPermissions,
-        tenantGrantFetchPending, dispatch
-      }
-      return <PermissionConsentType {...consentTypeProperties} />
+        item,
+        allPrincipalPermissions,
+        singlePrincipalPermissions,
+        tenantGrantFetchPending,
+        dispatch
+      };
+      return <PermissionConsentType {...consentTypeProperties} />;
     }
     return null;
-  }
+  };
 
   const handleRevoke = async (permission: IPermission): Promise<void> => {
     dispatch(revokeScopes(permission.value));
@@ -83,83 +78,62 @@ const PermissionItem = (props: PermissionItemProps): JSX.Element | null => {
   const ConsentButton = (): JSX.Element => {
     if (consented) {
       if (userHasRequiredPermissions()) {
-        return <PrimaryButton onClick={() => handleRevoke(item)}
-          styles={consentButtonStyles}
-        >
-          {translateMessage('Revoke')}
-        </PrimaryButton>
+        return (
+          <Button
+            appearance='primary'
+            onClick={() => handleRevoke(item)}
+            className={styles.button}
+          >
+            {translateMessage('Revoke')}
+          </Button>
+        );
       }
-      return <TooltipHost
-        content={translateMessage('You require the following permissions to revoke')}
-        id={hostId}
-        calloutProps={{ gapSpace: 0 }}
-      >
-        <DefaultButton
-          iconProps={buttonIcon}
-          allowDisabledFocus
-          disabled={true}
-          styles={consentButtonStyles}>
-          {translateMessage('Revoke')}
-        </DefaultButton>
-      </TooltipHost>
+      return (
+        <Tooltip content={translateMessage('You require the following permissions to revoke')} relationship='label'>
+          <Button appearance='primary' disabled className={styles.button}>
+            {translateMessage('Revoke')}
+          </Button>
+        </Tooltip>
+      );
     }
-    return <PrimaryButton onClick={() => handleConsent(item)}
-      styles={consentButtonStyles}
-    >
-      {translateMessage('Consent')}
-    </PrimaryButton>;
-  }
+    return (
+      <Button appearance='primary' onClick={() => handleConsent(item)} className={styles.button}>
+        {translateMessage('Consent')}
+      </Button>
+    );
+  };
+
 
   if (column) {
     const content = item[column.fieldName as keyof IPermission] as string;
     switch (column.key) {
-
     case 'value':
-      if (props.index === 0) {
-        return <div>
+      return (
+        <div>
           {content}
-          <TooltipHost
-            content={translateMessage('Least privileged permission')}
-            id={'leastPrivilegedPermissionsTooltipHost'}
-            calloutProps={{ gapSpace: 0 }}
-          >
-            <IconButton
-              iconProps={infoIcon}
-              styles={{
-                root: {
-                  position: 'relative',
-                  top: '4px'
-                }
-              }}
-            />
-          </TooltipHost>
+          {props.index === 0 && (
+            <Tooltip content={translateMessage('Least privileged permission')} relationship='label'>
+              <span className={styles.icon}><InfoRegular /></span>
+            </Tooltip>
+          )}
         </div>
-      }
-      return <div>{content}</div>
+      );
 
     case 'isAdmin':
-      return <div style={adminLabelStyles}>
-        <Label>{translateMessage(item.isAdmin ? 'Yes' : 'No')}</Label>
-      </div>;
+      return (
+        <div className={styles.adminLabel}>
+          <Label weight='semibold'>{translateMessage(item.isAdmin ? 'Yes' : 'No')}</Label>
+        </div>
+      );
 
     case 'consented':
       return <ConsentButton />;
 
     case 'consentDescription':
       return (
-        <>
-          <TooltipHost
-            content={item.consentDescription}
-            id={hostId}
-            calloutProps={{ gapSpace: 0 }}
-            styles={{
-              root: { display: 'block' }
-            }}
-          >
-            <span aria-labelledby={hostId}>
-              {item.consentDescription}
-            </span>
-          </TooltipHost></>
+        <Tooltip content={item.consentDescription} relationship='label'>
+          <span>{item.consentDescription}</span>
+        </Tooltip>
       );
 
     case 'consentType':
@@ -167,19 +141,13 @@ const PermissionItem = (props: PermissionItemProps): JSX.Element | null => {
 
     default:
       return (
-        <TooltipHost
-          content={content}
-          id={hostId}
-          calloutProps={{ gapSpace: 0 }}
-        >
-          <span aria-labelledby={hostId} style={{ fontSize: FontSizes.medium }}>
-            {content}
-          </span>
-        </TooltipHost>
+        <Tooltip content={content} relationship='label'>
+          <span>{content}</span>
+        </Tooltip>
       );
     }
   }
   return null;
-}
+};
 
 export default PermissionItem;
