+++ src/app/views/query-runner/request/permissions/columns.tsx
@@ -1,144 +1,128 @@
-import { IColumn, IIconProps, IconButton, getTheme } from '@fluentui/react';
-
-import { componentNames, telemetry } from '../../../../../telemetry';
+import {
+  createTableColumn,
+  TableColumnDefinition,
+  Tooltip,
+  Button
+} from '@fluentui/react-components';
+import { telemetry, componentNames } from '../../../../../telemetry';
 import { ADMIN_CONSENT_DOC_LINK, CONSENT_TYPE_DOC_LINK } from '../../../../services/graph-constants';
 import { translateMessage } from '../../../../utils/translate-messages';
-import { styles } from '../../query-input/auto-complete/suffix/suffix.styles';
-import { permissionStyles } from './Permission.styles';
+import { IPermission } from '../../../../../types/permissions';
+import { InfoRegular } from '@fluentui/react-icons';
+import PermissionItem from './PermissionItem';
+import permissionStyles from './Permission.styles';
 
 type source = 'panel' | 'tab';
 
 interface ColumnProps {
   source: source;
   tokenPresent?: boolean;
-  onColumnClicked?: (ev: React.MouseEvent<HTMLElement>, column: IColumn) => void;
 }
 
 const trackLinkClickedEvent = (link: string, componentName: string) => {
   telemetry.trackLinkClickEvent(link, componentName);
-}
+};
 
 const openExternalWebsite = (url: string) => {
   switch (url) {
   case 'Consent type':
     window.open(CONSENT_TYPE_DOC_LINK, '_blank');
-    trackLinkClickedEvent(CONSENT_TYPE_DOC_LINK, componentNames.CONSENT_TYPE_DOC_LINK)
+    trackLinkClickedEvent(CONSENT_TYPE_DOC_LINK, componentNames.CONSENT_TYPE_DOC_LINK);
     break;
   case 'Admin consent required':
     window.open(ADMIN_CONSENT_DOC_LINK, '_blank');
     trackLinkClickedEvent(ADMIN_CONSENT_DOC_LINK, componentNames.ADMIN_CONSENT_DOC_LINK);
     break;
   }
-}
+};
 
-const getColumns = ({ source, tokenPresent, onColumnClicked }: ColumnProps): IColumn[] => {
-
-  const theme = getTheme();
-  const { columnCellStyles, cellTitleStyles } = permissionStyles(theme);
-
-  const infoIcon: IIconProps = {
-    iconName: 'Info',
-    styles: cellTitleStyles
+const createRenderColumnHeader = (styles: ReturnType<typeof permissionStyles>) => {
+  const RenderColumnHeader = (headerText: string): JSX.Element => {
+    const tooltipMessage =
+    headerText === 'Admin consent required'
+      ? translateMessage('Administrator permission')
+      : translateMessage('Permission consent type');
+    return (
+      <div className={styles.headerContainer}>
+        <Tooltip
+          content={tooltipMessage}
+          relationship='label'>
+          <Button
+            icon={<InfoRegular />}
+            appearance="subtle"
+            size="small"
+            className={styles.iconButton}
+            aria-label={translateMessage(headerText)}
+            onClick={() => openExternalWebsite(headerText)}
+          />
+        </Tooltip>
+        <span className={styles.headerText}>{translateMessage(headerText)}</span>
+      </div>
+    );
   };
+  RenderColumnHeader.displayName = 'RenderColumnHeader';
+  return RenderColumnHeader;
+};
 
-  const renderColumnHeader = (headerText: string) => {
-    if (headerText === 'Status') {
-      return (
-        <span style={{ position: 'relative', top: '2px', left: '2px', flex: 1 }}>
-          {translateMessage('Status')}
-        </span>
-      )
-    }
-
-    return (<div style={{ textAlign: 'left', display: 'flex' }}>
-      <IconButton
-        iconProps={infoIcon}
-        className={styles.iconButton}
-        id={'buttonId'}
-        ariaLabel={translateMessage(headerText)}
-        onClick={() => openExternalWebsite(headerText)}
-        styles={{ root: { position: 'relative', left: '4px' } }}
-      >
-      </IconButton>
-      <span style={{ paddingTop: '4px' }}>
-        {translateMessage(headerText)}
-      </span>
-    </div>)
-  }
-
-  const columns: IColumn[] = [
-    {
-      key: 'value',
-      name: translateMessage('Permission'),
-      fieldName: 'value',
-      minWidth: 110,
-      maxWidth: 200,
-      columnActionsMode: 0,
-      isResizable: true
-    }
+const getColumns = ({ source, tokenPresent }: ColumnProps):
+ TableColumnDefinition<{item: IPermission; index: number }>[] => {
+  const styles = permissionStyles();
+  const renderColumnHeader = createRenderColumnHeader(styles);
+  const columns: TableColumnDefinition<{ item: IPermission; index: number }>[] = [
+    createTableColumn({
+      columnId: 'value',
+      renderHeaderCell: () => translateMessage('Permission'),
+      renderCell: ({ item, index }) => {
+        return <PermissionItem column={{ key: 'value', fieldName: 'value' }} index={index} item={item} />;
+      }
+    })
   ];
 
   if (source === 'tab') {
     columns.push(
-      {
-        key: 'consentDescription',
-        name: translateMessage('Description'),
-        fieldName: 'consentDescription',
-        minWidth: 300,
-        isMultiline: true,
-        columnActionsMode: 0,
-        targetWidthProportion: 1,
-        flexGrow: 1,
-        isResizable: true
-      }
+      createTableColumn({
+        columnId: 'consentDescription',
+        renderHeaderCell: () => translateMessage('Description'),
+        renderCell: ({item, index}) => (
+          <PermissionItem
+            column={{ key: 'consentDescription', fieldName: 'consentDescription' }}
+            index={index}
+            item={item} />
+        )
+      })
     );
   }
 
   columns.push(
-    {
-      key: 'isAdmin',
-      name: translateMessage('Admin consent required'),
-      fieldName: 'isAdmin',
-      minWidth: 187,
-      ariaLabel: translateMessage('Administrator permission'),
-      isMultiline: true,
-      headerClassName: 'permissionHeader',
-      styles: columnCellStyles,
-      onRenderHeader: () => renderColumnHeader('Admin consent required'),
-      targetWidthProportion: 1,
-      flexGrow: 1
-    }
+    createTableColumn({
+      columnId: 'isAdmin',
+      renderHeaderCell: () => renderColumnHeader('Admin consent required'),
+      renderCell: ({item, index}) => (
+        <PermissionItem column={{ key: 'isAdmin', fieldName: 'isAdmin' }} index={index} item={item} />
+      )
+    })
   );
 
   if (tokenPresent) {
     columns.push(
-      {
-        key: 'consented',
-        name: translateMessage('Status'),
-        isResizable: false,
-        fieldName: 'consented',
-        minWidth: 130,
-        onRenderHeader: () => renderColumnHeader('Status'),
-        styles: columnCellStyles,
-        targetWidthProportion: 1,
-        flexGrow: 1
-      },
-      {
-        key: 'consentType',
-        name: translateMessage('Consent type'),
-        isResizable: false,
-        fieldName: 'consentType',
-        minWidth: 130,
-        onRenderHeader: () => renderColumnHeader('Consent type'),
-        styles: columnCellStyles,
-        ariaLabel: translateMessage('Permission consent type'),
-        targetWidthProportion: 1,
-        flexGrow: 1
-      }
+      createTableColumn({
+        columnId: 'consented',
+        renderHeaderCell: () => translateMessage('Status'),
+        renderCell: ({item, index}) => (
+          <PermissionItem column={{ key: 'consented', fieldName: 'consented' }} index={index} item={item} />
+        )
+      }),
+      createTableColumn({
+        columnId: 'consentType',
+        renderHeaderCell: () => renderColumnHeader('Consent type'),
+        renderCell: ({item, index}) => (
+          <PermissionItem column={{ key: 'consentType', fieldName: 'consentType' }} index={index} item={item} />
+        )
+      })
     );
   }
+
   return columns;
-}
+};
 
 export { getColumns };
-
