+++ src/app/views/sidebar/resource-explorer/collection/CollectionPermissions.tsx
@@ -1,18 +1,19 @@
+import React, { FC, useEffect, useState } from 'react';
 import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHeader,
-  TableHeaderCell,
-  TableRow,
+  CounterBadge,
+  FlatTree,
+  FlatTreeItem,
+  TreeItemLayout,
+  TreeOpenChangeEvent,
+  TreeOpenChangeData,
+  TreeItemValue,
   Label,
+  Spinner,
   Link,
   makeStyles,
-  Spinner,
   MessageBar,
   MessageBarBody
 } from '@fluentui/react-components';
-import { FC, useEffect } from 'react';
 import { useAppSelector } from '../../../../../store';
 import { componentNames, telemetry } from '../../../../../telemetry';
 import { CollectionPermission } from '../../../../../types/resources';
@@ -21,6 +22,8 @@ import { useCollectionPermissions } from '../../../../services/hooks/useCollecti
 import { translateMessage } from '../../../../utils/translate-messages';
 import { downloadToLocal, trackDownload } from '../../../common/download';
 import CommonCollectionsPanel from './CommonCollectionsPanel';
+import { formatScopeLabel } from './collection.util';
+import { PERMS_SCOPE } from '../../../../services/graph-constants';
 
 const useStyles = makeStyles({
   centeredLabel: {
@@ -30,7 +33,7 @@ const useStyles = makeStyles({
     justifyContent: 'center',
     alignItems: 'center'
   },
-  tableContainer: {
+  treeContainer: {
     height: '80vh',
     overflowY: 'auto',
     overflowX: 'hidden'
@@ -43,18 +46,22 @@ const CollectionPermissions: FC<PopupsComponent<null>> = (props) => {
   const defaultCollection = collections ? collections.find(k => k.isDefault) : null;
   const paths = defaultCollection ? defaultCollection.paths : [];
 
-  const columns = [
-    {
-      key: 'value', name: translateMessage('Permissions'), fieldName: 'value',
-      minWidth: 300,
-      ariaLabel: translateMessage('Value')
+  const styles = useStyles();
+
+  const [openItems, setOpenItems] = useState<Set<TreeItemValue>>(new Set());
+
+  useEffect(() => {
+    if (paths.length > 0) {
+      getPermissions(paths);
     }
-  ];
+  }, [paths, getPermissions]);
 
   const downloadPermissions = (): void => {
     const filename = 'collection-permissions.json';
-    downloadToLocal(permissions, filename);
-    trackDownload(filename, componentNames.DOWNLOAD_COLLECTION_PERMISSIONS_BUTTON);
+    if (permissions) {
+      downloadToLocal(permissions, filename);
+      trackDownload(filename, componentNames.DOWNLOAD_COLLECTION_PERMISSIONS_BUTTON);
+    }
   };
 
   const handleTelemetryClick = (
@@ -66,13 +73,25 @@ const CollectionPermissions: FC<PopupsComponent<null>> = (props) => {
     );
   };
 
-  useEffect(() => {
-    if (paths.length > 0) {
-      getPermissions(paths)
-    }
-  }, [paths]);
+  const permissionsArray: CollectionPermission[] = [];
+  if (permissions) {
+    Object.keys(permissions).forEach((key) => {
+      permissionsArray.push(...permissions[key]);
+    });
+  }
 
-  const styles = useStyles();
+  const groupedPermissions = new Map<string, CollectionPermission[]>();
+  permissionsArray.forEach((p) => {
+    const groupKey = p.scopeType || 'unknown';
+    if (!groupedPermissions.has(groupKey)) {
+      groupedPermissions.set(groupKey, []);
+    }
+    groupedPermissions.get(groupKey)?.push(p);
+  });
+
+  const handleOpenChange = (_event: TreeOpenChangeEvent, data: TreeOpenChangeData) => {
+    setOpenItems(data.openItems);
+  };
 
   if (!isFetching && !permissions) {
     return (
@@ -83,55 +102,63 @@ const CollectionPermissions: FC<PopupsComponent<null>> = (props) => {
   }
 
   if (isFetching) {
-    return (
-      <Spinner label={translateMessage('Fetching permissions')} />
-    );
-  }
-
-  const permissionsArray: CollectionPermission[] = [];
-  if (permissions) {
-    Object.keys(permissions).forEach(key => {
-      permissionsArray.push(...permissions[key]);
-    });
+    return <Spinner label={translateMessage('Fetching permissions')} />;
   }
 
   return (
     <CommonCollectionsPanel
-      primaryButtonText='Download permissions'
+      primaryButtonText="Download permissions"
       primaryButtonAction={downloadPermissions}
       primaryButtonDisabled={!permissions}
       closePopup={props.dismissPopup}
     >
-      <MessageBar intent='info'>
+      <MessageBar intent="info">
         <MessageBarBody>
-          {translateMessage('list of permissions')}
+          {translateMessage('list of permissions')}{' '}
           <Link
-            target='_blank'
+            target="_blank"
             rel="noopener noreferrer"
             onClick={handleTelemetryClick}
-            href={'https://learn.microsoft.com/graph/permissions-reference?view=graph-rest-1.0'}
+            href="https://learn.microsoft.com/graph/permissions-reference?view=graph-rest-1.0"
           >
             {translateMessage('Microsoft Graph permissions reference')}
           </Link>
         </MessageBarBody>
       </MessageBar>
-      <div className={styles.tableContainer}>
-        <Table>
-          <TableHeader>
-            <TableRow>
-              {columns.map((column) => (
-                <TableHeaderCell key={column.key}>{column.name}</TableHeaderCell>
-              ))}
-            </TableRow>
-          </TableHeader>
-          <TableBody>
-            {permissionsArray.map((permission, index) => (
-              <TableRow key={index}>
-                <TableCell>{permission.value}</TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
+
+      <div className={styles.treeContainer}>
+        <FlatTree
+          openItems={openItems}
+          onOpenChange={handleOpenChange}
+          aria-label={translateMessage('Permissions')}
+        >
+          {[...groupedPermissions.entries()].map(([scopeType, perms]) => (
+            <React.Fragment key={scopeType}>
+              <FlatTreeItem value={scopeType} itemType="branch" aria-level={1}
+                aria-setsize={2}
+                aria-posinset={perms.length + 1}>
+                <TreeItemLayout aside={
+                  <CounterBadge count={perms.length} color="informative" />
+                }>{formatScopeLabel(scopeType as PERMS_SCOPE)}</TreeItemLayout>
+              </FlatTreeItem>
+              {perms.map((permission) => {
+                return (
+                  <FlatTreeItem
+                    key={permission.value}
+                    value={permission.value}
+                    parentValue={scopeType}
+                    itemType="leaf"
+                    aria-level={2}
+                    aria-posinset={perms.findIndex((p) => p.value === permission.value) + 1}
+                    aria-setsize={perms.length}
+                  >
+                    <TreeItemLayout>{permission.value}</TreeItemLayout>
+                  </FlatTreeItem>
+                );
+              })}
+            </React.Fragment>
+          ))}
+        </FlatTree>
       </div>
     </CommonCollectionsPanel>
   );
