+++ src/app/views/sidebar/resource-explorer/collection/EditCollectionPanel.tsx
@@ -31,8 +31,8 @@ const EditCollectionPanel: React.FC<EditCollectionPanelProps> = ({ closePopup })
   const items = collections && collections.length > 0 ? collections.find(k => k.isDefault)!.paths : [];
 
   const columns = [
-    { key: 'url', name: 'URL', fieldName: 'url', minWidth: 300, maxWidth: 1100, isResizable: true },
-    { key: 'scope', name: 'Scope', fieldName: 'scope', minWidth: 150, maxWidth: 200, isResizable: true }
+    { key: 'url', name: 'Select all', fieldName: 'url', minWidth: 300, maxWidth: 1100, isResizable: true },
+    { key: 'scope', name: '', fieldName: 'scope', minWidth: 150, maxWidth: 200, isResizable: true }
   ];
 
   const removeSelectedItems = () => {
